+++ tests/collector_async.rs
@@ -1,4 +1,4 @@
-#![cfg(feature = "std")]
+#![cfg(all(feature = "std", feature = "collector"))]
 
 use benchmark::*;